requirements.txt
attrs==20.3.0
bleach==3.2.1
botocore==1.23.52
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.3
dnspython==2.2.1
docutils==0.16
Flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.1.0
keyring==21.4.0
lightgbm==3.3.2
MarkupSafe==2.0.1
numpy==1.22.2
packaging==20.4
pandas==1.4.1
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
py-openaq==1.1.0
Pygments==2.7.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.1
readme-renderer==35.0
requests==2.25.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.3.6
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
soupsieve==2.2.1
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
twine==3.2.0
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==2.0.1
xgboost==1.5.2
